\name{initiate}
\alias{initiate}

\title{Initiate new 'cystiSim' object}

\description{
Intervention function that mimics the possible effects of human MDA.
}

\usage{
initiate(man, pig, ph2m, pl2m, m2p, e2p, age.coef = c(0, 0))
}

\arguments{
  \item{man}{.}
  \item{pig}{.}
  \item{ph2m}{Pig high to man transmission probability.}
  \item{pl2m}{Pig low to man transmission probability.}
  \item{m2p}{Man to pig transmission probability.}
  \item{e2p}{Environment to pig transmission probability.}
  \item{age.coef}{.}
}

\details{}

\author{\email{Brecht.Devleesschauwer@UGent.be}}

\seealso{}