\name{man_mbeya}
\alias{man_mbeya}

\docType{data}

\title{Mbeya human population}

\description{Default baseline dataframe.}

\usage{data("man_mbeya")}

\format{
  A data frame with 6000 observations on the following 7 variables.
  \describe{
    \item{\code{age}}{a numeric vector}
    \item{\code{sex}}{a factor with levels \code{female} \code{male}}
    \item{\code{taenia}}{a numeric vector}
    \item{\code{taenia_immature}}{a numeric vector}
    \item{\code{time_since_infection}}{a numeric vector}
    \item{\code{environment}}{a numeric vector}
    \item{\code{time_since_contamination}}{a numeric vector}
  }
}

\note{
For more details and examples, please visit the \pkg{cystiSim} Wiki pages on \url{https://github.com/brechtdv/cystiSim/wiki}.
}

\keyword{datasets}
