\name{pig_mbozi}
\alias{pig_mbozi}

\docType{data}

\title{Mbozi pig population}

\description{Default baseline dataframe.}

\usage{data("pig_mbozi")}

\format{
  A data frame with 498 observations on the following 8 variables.
  \describe{
    \item{\code{age}}{a numeric vector}
    \item{\code{cysti}}{a numeric vector}
    \item{\code{cysti_immature}}{a numeric vector}
    \item{\code{time_since_infection}}{a numeric vector}
    \item{\code{intensity}}{a factor with levels \code{0} \code{H} \code{L}}
    \item{\code{immunity}}{a numeric vector}
    \item{\code{time_since_vaccination}}{a logical vector}
    \item{\code{slaughtered}}{a numeric vector}
  }
}

\note{
For more details and examples, please visit the \pkg{cystiSim} Wiki pages on \url{https://github.com/brechtdv/cystiSim/wiki}.
}

\keyword{datasets}
