\name{baseline}
\alias{random_baseline_man}
\alias{random_baseline_pig}

\title{Random baseline populations}

\description{
Generate random baseline human and pig populations.
}

\usage{
random_baseline_man(n, p)
random_baseline_pig(n, p, p.high)
}

\arguments{
  \item{n}{Number of individuals to simulate.}
  \item{p}{Proportion of individuals infected with a mature parasite.}
  \item{p.high}{Proportion of infected pigs with high infection intensity.}
}

%\details{}

\author{\email{brechtdv@gmail.com}}

